Resetting functionality

Index

EnergyModelsRecedingHorizon.AbstractInitDataPath

EnergyModelsRecedingHorizon.AbstractPath

EnergyModelsRecedingHorizon.AbstractReset

EnergyModelsRecedingHorizon.AbstractSub

EnergyModelsRecedingHorizon.ElementPath

EnergyModelsRecedingHorizon.ElementReset

EnergyModelsRecedingHorizon.FutureValueSub

EnergyModelsRecedingHorizon.InitDataPath

EnergyModelsRecedingHorizon.InitReset

EnergyModelsRecedingHorizon.LinkSub

EnergyModelsRecedingHorizon.ModelSub

EnergyModelsRecedingHorizon.NodeSub

EnergyModelsRecedingHorizon.OperPath

EnergyModelsRecedingHorizon.OperReset

EnergyModelsRecedingHorizon.ProductSub

EnergyModelsRecedingHorizon.TimeWeightPath

EnergyModelsRecedingHorizon.TimeWeightReset

EnergyModelsRecedingHorizon.UpdateCase

EnergyModelsRecedingHorizon.ResetType

EnergyModelsRecedingHorizon.Substitution

EnergyModelsRecedingHorizon._ele_to_sub

EnergyModelsRecedingHorizon.get_sub_ele

EnergyModelsRecedingHorizon.get_sub_elements_vec

EnergyModelsRecedingHorizon.get_sub_model

EnergyModelsRecedingHorizon.get_sub_periods

EnergyModelsRecedingHorizon.get_sub_products

EnergyModelsRecedingHorizon.is_init_reset

EnergyModelsRecedingHorizon.model_key

EnergyModelsRecedingHorizon.original

EnergyModelsRecedingHorizon.resets

EnergyModelsRecedingHorizon.updated

Paths

Path are internal types that are used for the identification of a reset. They are used internally for dispatch.

EnergyModelsRecedingHorizon.AbstractPath — Type

abstract type AbstractPath

Supertype for identification of the type which is resetted by a lens.

EnergyModelsRecedingHorizon.OperPath — Type

struct OperPath <: AbstractPath

Internal type for paths pointing towards operational profiles.

EnergyModelsRecedingHorizon.ElementPath — Type

struct ElementPath <: AbstractPath

Internal type for paths pointing towards elements.

EnergyModelsRecedingHorizon.TimeWeightPath — Type

struct TimeWeightPath <: AbstractPath

Internal type for paths pointing towards the time weight of a future value.

EnergyModelsRecedingHorizon.AbstractInitDataPath — Type

abstract type AbstractInitDataPath <: AbstractPath

Supertype for identification of the AbstractPath types that are used for initial data. If you create a new AbstractInitData type, you must also create a new AbstractInitDataPath subtype

EnergyModelsRecedingHorizon.InitDataPath — Type

struct InitDataPath <: AbstractInitDataPath

Internal type for paths pointing towards InitData.

Fields

key::Symbol is the key in the InitData (and correspondingly in the JuMP model).

The following function is included for extracting the values from paths:

EnergyModelsRecedingHorizon.model_key — Method

model_key(idp::InitDataPath)

Returns the model key (field key) of InitDataPath idp.

Resetting new values

If you must reset a different type of value than outlined below, you must also create a new type. This is especially important when creating a new AbstractInitData.

Reset types

Reset types are introduced to differentiate on what must be reset in a given element. They are used internally for dispatch.

EnergyModelsRecedingHorizon.AbstractReset — Type

abstract type AbstractReset

Supertype for types resetting values in fields in the individual AbstractElements. The individual subtype is related to the chosen AbstractPath as outlined above.

New subtypes

We currently have reset types for elements (e.g., nodes, links, areas, or transmission), operational profiles, and initial data. The individual types are automatically deduced through the constructor ResetType.

If you require resetting different fields than the provided, you must include a new AbstractPath subtype, a new mutable composite type as subtype of AbstractReset, and a new method for the the constructor ResetType.

EnergyModelsRecedingHorizon.ElementReset — Type

mutable struct ElementReset <: AbstractReset

AbstractReset for resetting elements within another. The inner constructor is utilized for automatically creating the lens to the field path.

Inner constructor arguments

field_path::Vector is the path towards the field as identified through the function _find_update_paths.

x is the instance of a type for which the reset type is created.

Fields

lens::Union{PropertyLens,ComposedFunction} is the lens for resetting the field.

val::AbstractElement is the element in the field of the original instance. It is used for identifying the linked AbstractElement after it is reset.

EnergyModelsRecedingHorizon.OperReset — Type

mutable struct OperReset <: AbstractReset

AbstractReset for resetting operational profiles within an element. The inner constructor is utilized for automatically creating the lens to the field path.

Inner constructor arguments

field_path::Vector is the path towards the field as identified through the function _find_update_paths.

x is the instance of a type for which the reset type is created.

Fields

lens::Union{PropertyLens,ComposedFunction} is the lens for resetting the field.

var is the variable when using ParametricOptInterface.

val::OperationalProfile is the complete operational profile from the original element. The individual values are extracted from this profile in the receding horizon framework.

EnergyModelsRecedingHorizon.TimeWeightReset — Type

mutable struct TimeWeightReset <: AbstractReset

AbstractReset for resetting the time weight used in future value calculations.

Inner constructor arguments

field_path::Vector is the path towards the field as identified through the function _find_update_paths.

x is the instance of a type for which the reset type is created.

Fields

lens::Union{PropertyLens,ComposedFunction} is the lens for resetting the field.

var is the variable when using ParametricOptInterface.

val is the time weight that should be used in the analysis.

EnergyModelsRecedingHorizon.InitReset — Type

mutable struct InitReset{T} <: AbstractReset where {T<:AbstractInitDataPath}

AbstractReset for resetting initial data of an element. The inner constructor is utilized for automatically creating the lens to the field path.

Inner constructor arguments

field_path::Vector is the path towards the field as identified through the function _find_update_paths.

path::AbstractInitDataPath is the AbstractPath of the init data. It includes additional information that is utilized when resetting an element.

x is the instance of a type for which the reset type is created.

Fields

lens::Union{PropertyLens,ComposedFunction} is the lens for resetting the field.

path::AbstractInitDataPath is the AbstractPath of the init data. It includes additional information that is utilized when resetting an element.

var is the variable when using ParametricOptInterface.

val is the initial data value that should be used in the analysis.

EnergyModelsRecedingHorizon.ResetType — Function

ResetType(field_path::Vector, _::OperPath, x)
ResetType(field_path::Vector, _::ElementPath, x)
ResetType(field_path::Vector, _::TimeWeightPath, x)
ResetType(field_path::Vector, path::AbstractInitDataPath, x)

Constructor for AbstractReset types depending on their specified AbstractPath.

New types

If you create a new AbstractReset type (due to the creation of a new [AbstractPath] type), you must also create a new method for this function.

This is not necessary for a new subtype of AbstractInitDataPath.

The following function is included for identification purposes:

EnergyModelsRecedingHorizon.is_init_reset — Function

is_init_reset(rt::AbstractReset)

Function fo identifying whether the AbstractReset rt is of type InitReset.

Resetting new values

If you must reset a different type of value than outlined below, you must also create a new type. If you create a new AbstractPath, it is not neccesary to create a new type.

In this case, you must also create a new method for ResetType as this function is used internally.

Substitution types

Substitution types correspond to types that store information of the individual original instance, the updated instance, and all required resets of an object:

EnergyModelsRecedingHorizon.AbstractSub — Type

abstract type AbstractSub

Supertype for the creation of Substitution types.

Substitution types are used for simpler identification of the structures for substitutions.

New elements

If you include a new element, you must introduce a new AbstractSub type for the element.

EnergyModelsRecedingHorizon.ModelSub — Type

mutable struct ModelSub{T<:RecHorEnergyModel} <: AbstractSub

AbstractSub for RecHorEnergyModel.

Fields

new::T is the new instance after resetting its values.

org::T is the original instance before resetting its values.

resets::Vector{<:AbstractReset} are AbstractReset types for the given RecHorEnergyModel.

EnergyModelsRecedingHorizon.ProductSub — Type

mutable struct ProductSub{T<:Resource} <: AbstractSub

AbstractSub for Resources.

Fields

new::T is the new instance after resetting its values.

org::T is the original instance before resetting its values.

resets::Vector{<:AbstractReset} are AbstractReset types for the given Resource.

EnergyModelsRecedingHorizon.NodeSub — Type

mutable struct NodeSub{T<:EMB.Node} <: AbstractSub

AbstractSub for Nodes.

Fields

new::T is the new instance after resetting its values.

org::T is the original instance before resetting its values.

resets::Vector{<:AbstractReset} are AbstractReset types for the given Node.

EnergyModelsRecedingHorizon.LinkSub — Type

mutable struct LinkSub{T<:Link} <: AbstractSub

AbstractSub for Links.

Fields

new::T is the new instance after resetting its values.

org::T is the original instance before resetting its values.

resets::Vector{<:AbstractReset} are AbstractReset types for the given Link.

EnergyModelsRecedingHorizon.FutureValueSub — Type

mutable struct FutureValueSub{T<:FutureValue} <: AbstractSub

AbstractSub for FutureValues.

Fields

new::T is the new instance after resetting its values.

org::T is the original instance before resetting its values.

resets::Vector{<:AbstractReset} are AbstractReset types for the given FutureValue.

EnergyModelsRecedingHorizon.Substitution — Function

Substitution(x::T, resets::Vector{<:AbstractReset}) where {T}
Substitution(new::T, org::T, resets::Vector{<:AbstractReset}) where {T<:Resource}
Substitution(new::T, org::T, resets::Vector{<:AbstractReset}) where {T<:RecHorEnergyModel}
Substitution(new::T, org::T, resets::Vector{<:AbstractReset}) where {T<:AbstractElement}

Constructor for AbstractSub types depending on the specified type of the input.

The following function is included for extracting the values from substitution types:

EnergyModelsRecedingHorizon.resets — Function

resets(s::AbstractSub)

Returns the Vector{AbstractReset} of the AbstractSub s.

The following function is included for simplifying the overall code structure:

EnergyModelsRecedingHorizon._ele_to_sub — Function

_ele_to_sub(::Type{<:EMB.Node})
_ele_to_sub(::Type{<:Link})
_ele_to_sub(::Type{<:FutureValue})

Returns the subtype of AbstractSub for a given element.

New elements

If you include a new element, you must introduce a new method for this function.

Introducing new Elements

If you introduce new AbstractElements, you must also create a new subtype of AbstractSub, similar to NodeSub or LinkSub. The new subtype should have the same fields as the other subtypes.

In this case, you must also create a new method for _ele_to_sub

Update case type

All update information for updating/resetting the case dictionary and the model is stored within a single type:

EnergyModelsRecedingHorizon.UpdateCase — Type

mutable struct UpdateCase <: AbstractCase

Type including all information required for updating the individual fields of the elements in a receding horizon model.

This type follows in general the same structure as the Case type introduced in EnergyModelsBase in which the individual vectors of Resources or AbstractElements are replaced with the corresponding vectors of AbstractSub.

Fields

model::ModelSub is the substitution type for the RecHorEnergyModel.

opers::Dict is a dictionary for mapping the operational periods of the receding horizon problem to the operational periods of the full problem.

products::Vector{<:ProductSub} is a vector of substitution types for the individual Resources.

elements::elements::Vector{Vector} is a vector of vectors of substitution types for the individual AbstractElements.

The following functions are included for extracting the values from paths:

EnergyModelsRecedingHorizon.get_sub_model — Function

get_sub_model(𝒰::UpdateCase)

Returns the ModelSub type of UpdateCase 𝒰.

EnergyModelsRecedingHorizon.get_sub_periods — Function

get_sub_periods(𝒰::UpdateCase)

Returns the periods mapping dictionary of UpdateCase 𝒰.

EnergyModelsRecedingHorizon.get_sub_products — Function

get_sub_products(𝒰::UpdateCase)

Returns the products Vector of UpdateCase 𝒰.

EnergyModelsRecedingHorizon.get_sub_elements_vec — Function

get_sub_elements_vec(𝒰::UpdateCase)

Returns the elements Vector{Vector} of UpdateCase 𝒰.

EnergyModelsRecedingHorizon.get_sub_ele — Function

get_sub_ele(𝒮ᵛᵉᶜ::Vector{Vector}, x::Type{<:AbstractElement})
get_sub_ele(𝒰::UpdateCase, x::Type{<:AbstractElement})

Returns the Vector of AbstractSub corresponding to the type provided by x.

The function can be used both on an UpdateCase and the corresponding Vector{Vector}.

EnergyModelsRecedingHorizon.updated — Function

updated(𝒰::UpdateCase, x_org::AbstractElement)
updated(𝒰::UpdateCase, x_org::Resource)
updated(𝒮::Vector{<:AbstractSub}, x_org::AbstractElement)
updated(s::AbstractSub)

Returns the updated (resetted) instance of the original instance x_org for a given UpdateCase. It is used for mapping and replacing instances of the type in fields.

If the input is an AbstractSub, it returns the value of the field new.

EnergyModelsRecedingHorizon.original — Function

original(𝒰::UpdateCase, x_new::AbstractElement)
original(𝒰::UpdateCase, x_new::Resource)
original(𝒰::UpdateCase, x_new::TS.TimePeriod)
original(𝒮::Vector{<:AbstractSub}, x_new::AbstractElement)
original(s::AbstractSub)

Returns the original instance of the new (resetted) instance x_new for a given UpdateCase. It is used for results extraction.

If the input is an AbstractSub, it returns the value of the field org.