Types

Index

EnergyModelsRecedingHorizon.EnergyModelsRecedingHorizon

EnergyModelsRecedingHorizon.AbstractHorizons

EnergyModelsRecedingHorizon.RecHorEnergyModel

EnergyModelsRecedingHorizon.SingleHorizon

Types

EnergyModelsRecedingHorizon.EnergyModelsRecedingHorizon — Module

Main module for EnergyModelsRecedingHorizon, a package that extends EnergyModelsX with the implementation of a receding horizon framework for stress testing the invested energy system in an operational framework.

Model

EnergyModelsRecedingHorizon.RecHorEnergyModel — Type

abstract type RecHorEnergyModel <: EnergyModel end

Abstract type for receding horizon models.

Horizons

EnergyModelsRecedingHorizon.AbstractHorizons — Type

AbstractHorizons{T<:Real}

Abstract supertype for the individual composite types that can be used for passing the required parameters for a receding horizon optimization.

EnergyModelsRecedingHorizon.SingleHorizon — Type

SingleHorizon{T}

Type created as a result of iterating through an AbstractHorizons type. It stores the required information of each individual optimization.

Fields

id::Int64 is an integer corresponding to the number of the type from the iterator.

dur::Vector{T} is a vector of the duration of the operational period included in the horizon.

idx_optim::Vector{Int64} is a vector containing the indices used in the optimization horizon.

idx_impl::Vector{Int64} is a vector containing the indices used in the implementation horizon.